+++ coaster/utils.py
@@ -22,7 +22,8 @@ from unidecode import unidecode
 import html5lib
 import bleach
 import isoweek
-import base58
+
+from .shortuuid import suuid, encode as uuid2suuid, decode as suuid2uuid  # noqa
 
 if six.PY3:
     from html import unescape
@@ -142,27 +143,6 @@ def buid2uuid(value):
     UUID('33203dd2-f2ef-422f-aeb0-058d6f5f7089')
     """
     return uuid.UUID(bytes=urlsafe_b64decode(str(value + '==')))
-
-
-def suuid():
-    """
-    Return a ShortUUID using uuid1mc rendered in Base58
-    """
-    return base58.b58encode(uuid1mc().bytes)
-
-
-def uuid2suuid(value):
-    """
-    Render a UUID as a Base58 ShortUUID
-    """
-    return base58.b58encode(value.bytes)
-
-
-def suuid2uuid(value):
-    """
-    Convert a Base58 ShortUUID back into a UUID
-    """
-    return uuid.UUID(bytes=base58.b58decode(value))
 
 
 def newsecret():